refactor: MeetingMember 패키지 변경
