+++ src/main/java/com/mobble/mobbleserver/refactor/meeting/entity/Meeting.java
@@ -3,7 +3,7 @@ package com.mobble.mobbleserver.refactor.meeting.entity;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.meeting.MeetingErrorCode;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
-import com.mobble.mobbleserver.refactor.meetingMember.entity.MeetingMember;
+import com.mobble.mobbleserver.domain.meetingMember.MeetingMember;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
