+++ src/test/java/com/mobble/mobbleserver/refactor/meetingMember/entity/MeetingMemberTest.java
@@ -1,5 +1,6 @@
 package com.mobble.mobbleserver.refactor.meetingMember.entity;
 
+import com.mobble.mobbleserver.domain.meetingMember.MeetingMember;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.meeting.MeetingMemberErrorCode;
 import com.mobble.mobbleserver.refactor.club.core.entity.Club;
